+++ lib/phonelib.rb
@@ -4,6 +4,7 @@ module Phonelib
   autoload :Core, 'phonelib/core'
   autoload :Phone, 'phonelib/phone'
   autoload :PhoneAnalyzer, 'phonelib/phone_analyzer'
+  autoload :PhoneExtendedData, 'phonelib/phone_extended_data'
 
   extend Module.new {
     include Core