+++ lib/phonelib/phone.rb
@@ -2,10 +2,12 @@ module Phonelib
   # class for parsed phone number, includes validation and formatting methods
   class Phone
     # defining reader methods for class variables
-    attr_reader :original # original phone number passed for parsing
+    attr_reader :original  # original phone number passed for parsing
+    attr_reader :extension # phone extension passed for parsing after a number
 
     # including module that has all phone analyzing methods
     include Phonelib::PhoneAnalyzer
+    include Phonelib::PhoneExtendedData
 
     # class initialization method
     #
@@ -16,7 +18,8 @@ module Phonelib
     #   country (2 letters) like 'US', 'us' or :us for United States
     #
     def initialize(original, country = nil)
-      @original = original
+      @original, @extension = separate_extension(original)
+      @extension.gsub!(/[^0-9]/, '') if @extension
 
       if sanitized.empty?
         @data = {}
@@ -190,28 +193,13 @@ module Phonelib
       !valid_for_country?(country)
     end
 
-    # Returns geo name of parsed phone number or nil if number is invalid or
-    # there is no geo name specified in db for this number
-    def geo_name
-      get_ext_name Phonelib::Core::EXT_GEO_NAMES,
-                   Phonelib::Core::EXT_GEO_NAME_KEY
-    end
-
-    # Returns timezone of parsed phone number or nil if number is invalid or
-    # there is no timezone specified in db for this number
-    def timezone
-      get_ext_name Phonelib::Core::EXT_TIMEZONES,
-                   Phonelib::Core::EXT_TIMEZONE_KEY
-    end
-
-    # Returns carrier of parsed phone number or nil if number is invalid or
-    # there is no carrier specified in db for this number
-    def carrier
-      get_ext_name Phonelib::Core::EXT_CARRIERS,
-                   Phonelib::Core::EXT_CARRIER_KEY
-    end
-
     private
+
+    # extracts extension from passed phone number if provided
+    def separate_extension(original)
+      splitted = (original || '').split /[;#]/
+      [splitted.first, splitted[1..-1] && splitted[1..-1].join]
+    end
 
     # get main country for code among provided countries
     def get_main_country(countries_array)
@@ -220,49 +208,10 @@ module Phonelib
       end || countries_array.first
     end
 
-    # get name from extended phone data by keys
-    #
-    # ==== Attributes
-    #
-    # * +name_key+ - names array key from extended data hash
-    # * +id_key+   - parameter id key in resolved extended data for number
-    #
-    def get_ext_name(names_key, id_key)
-      if ext_data[id_key] > 0
-        res = Phonelib.phone_ext_data[names_key][ext_data[id_key]]
-        res.size == 1 ? res.first : res
-      end
-    end
-
-    # returns extended data ids for current number
-    def ext_data
-      return @ext_data if @ext_data
-
-      ext_keys = [
-          Phonelib::Core::EXT_GEO_NAME_KEY,
-          Phonelib::Core::EXT_TIMEZONE_KEY,
-          Phonelib::Core::EXT_CARRIER_KEY
-      ]
-      result = {}
-      ext_keys.each { |key| result[key] = 0 }
-
-      return result unless possible?
-
-      drill = Phonelib.phone_ext_data[Phonelib::Core::EXT_PREFIXES]
-
-      e164.gsub('+', '').each_char do |num|
-        drill = drill[num.to_i] || break
-
-        ext_keys.each do |key|
-          result[key] = drill[key] if drill[key]
-        end
-      end
-
-      @ext_data = result
-    end
-
     # Get needable data for formatting phone as national number
     def get_formatting_data
+      return @formatting_data if @formatting_data
+
       format = @data[country][:format]
       prefix = @data[country][Core::NATIONAL_PREFIX]
       rule = (format[Core::NATIONAL_PREFIX_RULE] ||
@@ -274,7 +223,8 @@ module Phonelib
       # add space to format groups, change first group to rule,
       format_string = format[:format].gsub(/(\d)\$/, '\\1 $').gsub('$1', rule)
 
-      [@national_number.match(/#{format[Core::PATTERN]}/), format_string]
+      @formatting_data =
+          [@national_number.match(/#{format[Core::PATTERN]}/), format_string]
     end
   end
 end